# src/internal/docs/swagger.yaml
basePath: /
definitions:
  models.AdminInput:
    properties:
      access:
        type: string
      userId:
        type: string
    type: object
  models.AdminOutput:
    properties:
      access:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/models.UserOutput'
    type: object
...
  models.Err:
    properties:
      description:
        description: the error description for the developer, aka err.Error()
        type: string
      message:
        description: the displayed error text for the user
        type: string
      name:
        type: string
      status:
        type: integer
    type: object
...
  models.Response:
    properties:
      data: {}
      timestamp:
        type: string
    type: object
...
  models.UserOutput:
    properties:
      _id:
        type: string
      active:
        type: boolean
      createdAt:
        type: string
      username:
        type: string
    type: object
host: 127.0.0.1:3000
info:
  contact:
    email: ??
    name: Tovah Parnes
  description: A server for BirdAI API, for managing users, admin, birds, posts and
    more.
  license:
    name: MIT License
    url: https://opensource.org/license/mit/
  termsOfService: http://swagger.io/terms/
  title: BirdAI API
  version: "2.0"
paths:
  /admins/:
    post:
      consumes:
      - application/json
      description: Create a new admin
      parameters:
      - description: admin
        in: body
        name: set
        required: true
        schema:
          $ref: '#/definitions/models.AdminInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Create a new admin
      tags:
      - Admins
  /admins/{id}:
    delete:
      consumes:
      - application/json
      description: Delete given admin
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Delete given admin
      tags:
      - Admins
    get:
      consumes:
      - application/json
      description: Get admin by ID
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AdminOutput'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "410":
          description: Gone
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get admin by ID
      tags:
      - Admins
    patch:
      consumes:
      - application/json
      description: Update given admin
      parameters:
      - description: admin ID
        in: path
        name: id
        required: true
        type: string
      - description: admin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Update given admin
      tags:
      - Admins
  /admins/list:
    get:
      consumes:
      - application/json
      description: List all admins of a specified set
      parameters:
      - description: Set of admins
        in: query
        name: set
        type: integer
      - description: Search parameter for admin
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AdminOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: List all admins of a specified set
      tags:
      - Admins
  /admins/me:
    get:
      consumes:
      - application/json
      description: Get current admin
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AdminOutput'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "403":
          description: Forbidden
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "410":
          description: Gone
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
        "503":
          description: Service Unavailable
          schema:
            allOf:
            - $ref: '#/definitions/models.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Err'
                  type: array
              type: object
      security:
      - Bearer: []
      summary: Get current admin
      tags:
      - Admins
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: '>- Enter the token with the `Bearer: ` prefix, e.g. "Bearer eyJhbGciOiJIUzI1NiJ9.e30.DXCaKJxPJq8JhXOZIN4yWdBwJ4sraVbwH9imDCHZPaA".'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"